+++ src/framework/piral-base/src/events.test.ts
@@ -4,31 +4,35 @@
 import { describe, it, expect, vitest } from 'vitest';
 import { createListener } from './events';
 
+function nextCycle(time = 0) {
+  return new Promise<void>(resolve => setTimeout(resolve, 0));
+}
+
 describe('Events Module', () => {
-  it('add and emit event', () => {
+  it('add and emit event', async () => {
     const events = createListener(undefined);
     const mockCallback = vitest.fn();
     events.on('init', mockCallback);
     events.emit('init', undefined);
 
-    setTimeout(() => {
-      expect(mockCallback).toHaveBeenCalledTimes(1);
-    }, 1);
+    await nextCycle(10);
+    
+    expect(mockCallback).toHaveBeenCalledTimes(1);
   });
 
-  it('does only react to self events when different states', () => {
+  it('does only react to self events when different states', async () => {
     const events1 = createListener({});
     const events2 = createListener({});
     const mockCallback = vitest.fn();
     events1.on('init', mockCallback);
     events2.emit('init', undefined);
 
-    setTimeout(() => {
-      expect(mockCallback).toHaveBeenCalledTimes(0);
-    }, 1);
+    await nextCycle(10);
+
+    expect(mockCallback).toHaveBeenCalledTimes(0);
   });
 
-  it('does only react to self events when same state', () => {
+  it('does only react to self events when same state', async () => {
     const state = {};
     const events1 = createListener(state);
     const events2 = createListener(state);
@@ -36,9 +40,9 @@ describe('Events Module', () => {
     events1.on('init', mockCallback);
     events2.emit('init', undefined);
 
-    setTimeout(() => {
-      expect(mockCallback).toHaveBeenCalledTimes(1);
-    }, 1);
+    await nextCycle(10);
+
+    expect(mockCallback).toHaveBeenCalledTimes(1);
   });
 
   it('emit on empty event should be fine', () => {
@@ -51,7 +55,7 @@ describe('Events Module', () => {
     events.off('init', vitest.fn());
   });
 
-  it('should not be possible to emit after event removed', () => {
+  it('should not be possible to emit after event removed', async () => {
     const events = createListener(undefined);
     const mockCallback = vitest.fn();
     events.on('init', mockCallback);
@@ -59,8 +63,8 @@ describe('Events Module', () => {
     events.off('init', mockCallback);
     events.emit('init', undefined);
 
-    setTimeout(() => {
-      expect(mockCallback).toHaveBeenCalledTimes(1);
-    }, 1);
+    await nextCycle(10);
+    
+    expect(mockCallback).toHaveBeenCalledTimes(1);
   });
 });