+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Piral Changelog
 
+## 1.4.1 (tbd)
+
+- Fixed issue with pilet injector for certain apps in monorepos
+- Added support for middleware functions in `piral-fetch` (#645)
+
 ## 1.4.0 (December 12, 2023)
 
 - Fixed issue when target tarball is part of tarball content in `pilet pack` (#642)